refactoring classes MealRepository and InMemoryMealRepositoryImpl

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,5 +1,6 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.slf4j.Logger;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
@@ -8,41 +9,60 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+
+import static org.slf4j.LoggerFactory.getLogger;
 
 /**
  * GKislin
  * 15.09.2015.
  */
 public class InMemoryMealRepositoryImpl implements MealRepository {
+    private static final Logger LOG = getLogger(InMemoryMealRepositoryImpl.class);
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(m -> save(m, m.getUser().getId()));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
+        if (meal.getUser().getId() != userId) {
+            return null;
+        }
+
+        String save = meal.isNew() ? "save " : "update ";
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
         }
+        LOG.info(save + meal);
         repository.put(meal.getId(), meal);
         return meal;
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public void delete(int id, int userId) {
+        LOG.info("delete " + id);
+        if (get(id, userId) != null) {
+            repository.remove(id);
+        }
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        LOG.info("get " + id);
+        Meal meal = repository.get(id);
+        return meal != null && meal.getUser().getId() == userId ? meal : null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(int userId) {
+        LOG.info("getAll with userId " + userId);
+        return repository.values().stream()
+                .filter(m -> m.getUser().getId() == userId)
+                .sorted((m1, m2) -> m2.getDateTime().compareTo(m1.getDateTime()))
+                .collect(Collectors.toList());
     }
 }
